+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Populate a value in Source field when View/Edit Inventory Holdings Record. Refs UIIN-1548.
 * Fix instance duplication when child or parent records are present. Fixes UIIN-1562.
 * Add bound-with icons and item detail view header note. Refs UIIN-1522, UIIN-1523, UIIN-1524.
+* Retrieve and display select Piece information on Holding. Refs UIIN-1502.
 
 
 ## [7.1.2](https://github.com/folio-org/ui-inventory/tree/v7.1.2) (2021-07-14)